pages/requirements.txt
streamlit
tensorflow
h5py
numpy
Pillow
opencv-python
python-dotenv
gpt4all
streamlit-js-eval
requests
